Organize indexer (file) utilities
